README.md: Remove $ and # from sh examples
Github's formatter treats lines beginning with # as comments, and the
'# #' is hard to read as text. Few users these days probably recognize
the importantance of $ vs # and the docs didn't really do a good job of
swapping between them anyway. It is probably better to just prefix
ddrescue invocations with sudo these days, but I'm not going to bother
at this time as I'd need to verify there wouldn't be common permission
problems.